split bind into three files and copy custom bindings from appsdk
